Changing RecordEncoder to generic Encoder
RecordEncoder doesn't actually do anything record-specific and we have other
objects (image metadata) that we want to encode with JSON. A normal JSON
encoder can't encode datetimes, but we can just make RecordEncoder a generic
Encoder that encodes datetimes as ISO-8601.

Also makes the small change that timestamps are assumed to be in UTC time
zone, related to the changes in the backend:

* https://github.com/JeetShetty/GreenPiThumb/pull/101
* https://github.com/JeetShetty/GreenPiThumb/pull/109

Changes the serialized format to append a Z (representing UTC time zone) instead
of an offset from UTC in hours.